backend/requirements.txt
requests==2.26.0
Django==3.2
djangorestframework==3.12.4
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
django-filter==22.1
gunicorn==20.1.0
python-dotenv==0.19.0
psycopg2-binary==2.9.3
django-cors-headers==4.1.0
djoser==2.1.0
django-colorfield==0.9.0
drf_yasg==1.21.4